refactor: Move domain and repository files outside internal
