+++ go/services/users/domain/repository/userRepository.go
@@ -1,7 +1,7 @@
 package repository
 
 import (
-	"GROWTHSPHERE/services/users/internal/domain/entity"
+	"GROWTHSPHERE/services/users/domain/entity"
 	"context"
 )
